Preserve field verifier result width
The save verifier returns a boolean byte, but its reconstructed ABI exposed a full word and changed the caller's branch. Record the target return width at the definition.

Constraint: The helper returns only zero or one and has a single active caller.

Rejected: Mask only at the callsite | that duplicates an ABI fact instead of fixing its source.

Confidence: high

Scope-risk: narrow

Tested: Forced field candidate rebuild; fn_801CF7E4 improved from 59.19% to 59.69%; git diff --check.

+++ src/game/field_range_801CB180.c
@@ -704,7 +704,7 @@ void fn_801CBBAC(u8 digest[20], const u8* input, u32 length)
     fn_801CBF64(digest, &context);
 }
 
-u32 fn_801CBCDC(u8* data, u32 size, const u32 expected[5], u32 offset)
+u8 fn_801CBCDC(u8* data, u32 size, const u32 expected[5], u32 offset)
 {
     u32* savedDigest;
     u32* digest;
